Rename "pool" package to "slot"
