+++ src/tes/worker/slot/helpers.go
@@ -1,4 +1,4 @@
-package pool
+package slot
 
 import (
 	pbe "tes/ga4gh"
