+++ src/tes/worker/slot/idle_timeout.go
@@ -1,4 +1,4 @@
-package pool
+package slot
 
 import "time"
