+++ src/tes/worker/slot/slot.go
@@ -1,4 +1,4 @@
-package pool
+package slot
 
 import (
 	"fmt"
